+++ Beleg/Beleg/Model.java
@@ -1,9 +1,8 @@
-package Beleg.Beleg;
+package Beleg;
 
 import java.awt.*;
 
 public class Model {
-		
 		  View v;
 		  boolean farbe = true;
 		  int threads = 10;
